+++ lib/src/utils/utils.module.dart
@@ -1,1 +1,2 @@
 export 'fns.util.dart';
+export 'extensions.util.dart';